task classes: move into dedicated task package
Task, Todo, Deadline, and Event classes are located in the root
package alongside unrelated classes.

Having task-related classes in the root package makes the codebase
less organized and harder to navigate as the number of classes grows.

Let's move Task, Todo, Deadline, and Event into a new `task` package
to improve code organization and maintainability. This change groups
related classes together, making the project structure clearer and
future extensions (e.g., adding new task types) easier to manage.

Imports in other classes are updated accordingly to reflect the new
locations.
