+++ src/main/java/capy/Storage.java
@@ -1,12 +1,15 @@
 package capy;
 
+import capy.task.Deadline;
+import capy.task.Event;
+import capy.task.Task;
+import capy.task.Todo;
+
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Scanner;
 import java.io.*;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 /**
  * Storage handles loading and saving tasks to the file system.
