+++ src/main/java/capy/TaskList.java
@@ -1,5 +1,7 @@
 package capy;
 
+import capy.task.Task;
+
 import java.util.ArrayList;
 import java.util.List;
