+++ src/main/java/capy/Ui.java
@@ -1,5 +1,7 @@
 package capy;
 
+import capy.task.Task;
+
 import java.util.List;
 import java.util.Scanner;
