+++ src/main/java/capy/task/Deadline.java
@@ -1,4 +1,4 @@
-package capy;
+package capy.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
