+++ src/test/java/capy/DeadlineTest.java
@@ -1,5 +1,6 @@
 package capy;
 
+import capy.task.Deadline;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
