+++ src/test/java/capy/TaskListTest.java
@@ -1,5 +1,6 @@
 package capy;
 
+import capy.task.Task;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
